requirements.txt
numpy==1.26.4
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
scipy==1.13.0
bcrypt==4.1.2
python-jose==3.3.0
pydantic[email]
